+++ carflux_frontend/js/utils.js
@@ -1,0 +1,6 @@
+ export function formatarPreco(valor) {
+    return parseFloat(valor).toLocaleString('pt-BR', {
+      style: 'currency',
+      currency: 'BRL'
+    });
+  }
