+++ carflux_frontend/pages/apagarRegistro.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="../assets/css/style.css">
   <title>Sistema de Veículos</title>
   
 </head>
@@ -17,7 +17,7 @@
   </div>
 
   <div class="main-content">
-    <h1>Cadastro de Veículo</h1>
+    <h1>Apagar Registro</h1>
 
     <form id="formVeiculo">
       <label for="marca">Marca:</label>
@@ -54,11 +54,11 @@
       <label for="preco">Preço (R$):</label>
       <input type="number" step="0.01" id="preco" name="preco" required>
 
-      <button type="submit">Cadastrar Veículo</button>
+      <button type="submit">Apagar Registro</button>
     </form>
   </div>
 
-  <script type="module" src="js/cadastrarVeiculo.js" ></script>
+  <script type="module" src="../js/cadastrarVeiculo.js" ></script>
 
  
 </body>
