+++ carflux_frontend/pages/listarVeiculos.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.6/dist/css/bootstrap.min.css">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
-  <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="../assets/css/style.css">
   <title>CAR FLUX</title>
   
 </head>
@@ -104,7 +104,8 @@
   </div>
 </div>
 
-<script src="js/listarVeiculos.js" ></script>
+<script type="module" src="../js/listarVeiculos.js" ></script>
+<script type="module" src="../js/components/modal.js"></script>
